+++ warp/tests/unittest_suites.py
@@ -88,6 +88,7 @@ def default_suite(test_loader: unittest.TestLoader = unittest.defaultTestLoader)
     from warp.tests.cuda.test_async import TestAsync
     from warp.tests.cuda.test_clang_cuda import TestClangCUDA
     from warp.tests.cuda.test_cuda_arch_suffix import TestCudaArchSuffix
+    from warp.tests.cuda.test_graph_specialize import TestGraphSpecialize
     from warp.tests.cuda.test_mempool import TestMempool
     from warp.tests.cuda.test_multigpu import TestMultiGPU
     from warp.tests.cuda.test_peer import TestPeer
@@ -215,6 +216,7 @@ def default_suite(test_loader: unittest.TestLoader = unittest.defaultTestLoader)
         TestArray,
         TestArrayReduce,
         TestAsync,
+        TestGraphSpecialize,
         TestAtomic,
         TestAtomicBitwise,
         TestAtomicCAS,